requirements.txt
streamlit==1.22.0
openai-whisper==1.0.1
nltk==3.8.1
python-dotenv==1.0.0
langchain==0.0.7
python-docx==0.8.11
webvtt-py==0.5.0
pydub==0.25.1